Fix all compiler warnings on Linux and Windows targets
- Silenced unused parameter warning in wasapi_capabilities_default_device_name.
- Resolved signedness ternary operator warnings in dsp_engine_core.c and pipeline.c by adding explicit size_t casts.
- Avoided GCC format/stringop truncation warnings in dsp_engine.c and dsp_engine_core.c by swapping snprintf and strncpy copies with memcpy/strncpy length-terminated logic.
- Wrap unused vDSP wrappers under defined(__APPLE__) in vdsp_complex_dft.c and vdsp_real_fft.c.

TAG=agy
CONV=af905aae-0b3f-4101-9a0b-dea0616fa4f8

+++ Sources/CDSP/Pipeline/pipeline.c
@@ -68,7 +68,7 @@ pipeline_t* pipeline_create(const dsp_config_t* config, processing_parameters_t*
         return NULL;
     }
 
-    pipeline->frames_per_chunk = explicit_chunk_size > 0 ? explicit_chunk_size : config->devices.chunksize;
+    pipeline->frames_per_chunk = explicit_chunk_size > 0 ? explicit_chunk_size : (size_t)config->devices.chunksize;
     pipeline->rate = config->devices.samplerate;
     pipeline->expected_in_channels = config->devices.capture.channels;
